Add new data structures for representing Trees: Tree, TreeMap and Path
- Add new predicates for Object and String that can be used in Preconditions and Assertions.
- Remove some useless @IngoreWarning annotations
- Add Unit Tests for the new data structures

+++ commons-core/src/main/java/org/atlanmod/commons/mutable/MutableInteger.java
@@ -125,12 +125,12 @@ public class MutableInteger extends Number implements Comparable<MutableInteger>
 
     @Override
     public Integer getValue() {
-        return Integer.valueOf(this.value);
+        return this.value;
     }
 
     @Override
     public void setValue(Integer newValue) {
-        this.value = newValue.intValue();
+        this.value = newValue;
     }
 
     @Override
